feat(color)!: remove Colorf interface
make Colorf as a class.

it is not planned for user to implement this
interface, so a class is used instead.

of course, some possibility for different color
models is still present, but it is not worth it.

to keep the library design more minimal and mature,
i decided to make this type a class.