refactor(geom)!: remove Pos2f/Pos2i
probably, the division for two separate classes
(one with setters and one without them) is an
overkill. in the context of a rasterization
process, the "constant" points are very
inefficient don't make sense.

also, the Triangle interface was using Pos,
and not Point. why? because they have
all the methods i need for the process - only
getters. but if a user wants to also use this
triangle? only getters are not so good in this
situation, as i believe.

so, for simplicity and other reasons, Pos was
removed.