refactor(triangle/geom)!: Polygon3
Polygon3 is the new name for the old DynamicTriangle.

firstly, the implementation of the Triangle
is not needed internally in the library. but i
need to implement it for the sake of providing
standard implementations for every interface and
my own testing.

secondly, "static" implementation doesn't
make sense. so only one implementation for the
Triangle interface remains.